+++ apps/web/app/admin/page.tsx
@@ -1,5 +1,5 @@
-import { Admin } from "../../components/pages";
 import { AppbarClient } from "../../components/AppbarClient";
+import { Admin } from "../../screens/Admin";
 
 export default async function AdminPage() {
   return (
