--- apps/web/components/pages/index.ts
@@ -1,2 +1,0 @@
-export * from "./Signup";
-export * from "./Admin";
